Improve runtime module graph for HMR
Focusing on edits to queries.ts. To my surprise Codex's suggestion to extract the class component proved effective. React Refresh has lots of odd limitations. 🤷

The dev server needs restarting for the change to become effective.
